corrections

+++ src/n1exercici5/PiValue.java
@@ -2,5 +2,5 @@ package n1exercici5;
 
 @FunctionalInterface
 public interface PiValue {
-    float getPiValue();
+    double getPiValue();
 }
